final with main
 Develop (in a team of 2) an application that reads two files for cities and countries
and store each in a List.
 Create a map that uses the country code as keys and a list of cities as the value for
each country.
 For a given country code (from Console) sort the cities according to the population
 Provide classes or methods to get the following:
 Get a List of countries population
 Get the average countries population
 Get the maximum countries population
 Highest population city of each country
 Highest population capital

+++ 01.java development SE/ CitiesAndCountries/src/Country.java
@@ -3,12 +3,12 @@ public class Country {
     private String countryCode;
     private String countryName;
     private String continent;
-    private long population;
+    private double population;
     private double surfaceArea;
     private double gnp;
     private int capitalId;
 
-    public Country(String countryCode, String countryName, String continent, long population, double surfaceArea, double gnp, int capitalId) {
+    public Country(String countryCode, String countryName, String continent, double population, double surfaceArea, double gnp, int capitalId) {
         this.countryCode = countryCode;
         this.countryName = countryName;
         this.continent = continent;
@@ -30,7 +30,7 @@ public class Country {
         return continent;
     }
 
-    public long getPopulation() {
+    public double getPopulation() {
         return population;
     }
